refactor: useWMS or useWMTS hooks are made from the same function
